requirements.txt
streamlit
PyMuPDF
openai
python-dotenv
langchain
langchain-huggingface
langchain-community
pinecone-client
streamlit-chat
